split realtime single day occupancies to a different endpoint
