+++ frontend/src/employee-frontend/components/unit/queries.ts
@@ -31,7 +31,8 @@ import {
 } from '../../generated/api-clients/daycare'
 import {
   getOccupancyPeriodsSpeculated,
-  getUnitOccupancies
+  getUnitOccupancies,
+  getUnitRealizedOccupanciesForDay
 } from '../../generated/api-clients/occupancy'
 import {
   createGroupPlacement,
@@ -57,6 +58,11 @@ export const queryKeys = createQueryKeys('unit', {
     to: LocalDate,
     groupId: UUID | null
   ) => ['unitOccupancies', unitId, { from, to }, groupId],
+  unitRealizedOccupanciesForDay: (
+    unitId: UUID,
+    date: LocalDate,
+    groupId: UUID | null
+  ) => ['unitRealizedOccupanciesForDay', unitId, date, groupId],
   unitApplications: (unitId: UUID) => ['unitApplications', unitId],
   unitGroups: (unitId: UUID) => ['unitGroups', unitId],
   unitSpeculatedOccupancyRates: (
@@ -104,6 +110,12 @@ export const unitOccupanciesQuery = query({
   api: getUnitOccupancies,
   queryKey: ({ unitId, from, to, groupId }) =>
     queryKeys.unitOccupancies(unitId, from, to, groupId ?? null)
+})
+
+export const unitRealizedOccupanciesForDayQuery = query({
+  api: getUnitRealizedOccupanciesForDay,
+  queryKey: ({ unitId, date, groupId }) =>
+    queryKeys.unitRealizedOccupanciesForDay(unitId, date, groupId ?? null)
 })
 
 export const unitApplicationsQuery = query({
